integrate TransformerModel to LitTransformer
