+++ src/models/models.py
@@ -26,54 +26,3 @@ class PositionalEncoding(nn.Module):
         """
         seq_len = x.size(0)
         return x + self.pe[:seq_len]
-
-
-class TransformerModel(nn.Module):
-    def __init__(
-        self,
-        src_vocab_size,
-        tgt_vocab_size,
-        src_max_len,
-        tgt_max_len,
-        d_model=512,
-        nhead=8,
-        num_encoder_layers=6,
-        num_decoder_layers=6,
-        dim_feedforward=2048,
-        dropout=0.1,
-    ):
-        super(TransformerModel, self).__init__()
-        self.src_embedding = nn.Embedding(src_vocab_size, d_model)
-        self.tgt_embedding = nn.Embedding(tgt_vocab_size, d_model)
-        self.transformer = nn.Transformer(
-            d_model,
-            nhead,
-            num_encoder_layers,
-            num_decoder_layers,
-            dim_feedforward,
-            dropout,
-        )
-        self.fc_out = nn.Linear(d_model, tgt_vocab_size)
-        self.d_model = d_model
-        self.src_pos_enc = PositionalEncoding(d_model, src_max_len)
-        self.tgt_pos_enc = PositionalEncoding(d_model, tgt_max_len)
-
-    def forward(self, src, tgt):
-        src = self.src_embedding(src) * torch.sqrt(
-            torch.tensor(self.d_model, dtype=torch.float32)
-        )
-        tgt = self.tgt_embedding(tgt) * torch.sqrt(
-            torch.tensor(self.d_model, dtype=torch.float32)
-        )
-        src = src.permute(1, 0, 2)  # (S, N, E)
-        tgt = tgt.permute(1, 0, 2)  # (T, N, E)
-        src = src + self.src_pos_enc(src)
-        tgt = tgt + self.tgt_pos_enc(tgt)
-        tgt_mask = nn.Transformer.generate_square_subsequent_mask(tgt.size(0))
-        if self.training:
-            output = self.transformer(src, tgt, tgt_mask=tgt_mask)
-        else:
-            self.transformer.eval()
-            output = self.transformer(src, tgt, src_mask=None, tgt_mask=None)
-        output = self.fc_out(output)
-        return output
